changed routes to views, moved things around, cleaned things up
