+++ src/components/RouteTabs.jsx
@@ -8,23 +8,23 @@ import { AttachMoney as MoneyIcon } from '@material-ui/icons';
 import { Store as StoreIcon } from '@material-ui/icons';
 import { AddLocation as AddLocationIcon } from '@material-ui/icons';
 
-const Body = lazy(() => import('./routes/Body'))
-const Store = lazy(() => import('./routes/DAppStore'))
-const GetHydro = lazy(() => import('./routes/GetHydro'))
-const FinalizeClaim = lazy(() => import('./routes/FinalizeClaim'))
-const Faq = lazy(() => import('../pages/FAQ'))
-const Stats = lazy(() => import('../pages/Stats'))
-const Audits = lazy(() => import('../pages/Audits'))
-const Terms = lazy(() => import('../pages/Terms'))
-const Privacy = lazy(() => import('../pages/Privacy'))
-const About = lazy(() => import('../pages/About'))
-const Contact = lazy(() => import('../pages/Contact'))
-const SubmitDApp = lazy(() => import('../pages/SubmitDApp'))
-const ManageIdentity = lazy(() => import('../pages/ManageYourIdentity'))
-const ManageWallet = lazy(() => import('../pages/DAppStoreWallet'))
-const ManageDApps = lazy(() => import('../pages/DAppsAdded'))
-const DAppCategories = lazy(() => import('../pages/DAppCategories'))
-const DAppDetails = lazy(() => import('../pages/DAppDetails'))
+const Body = lazy(() => import('./Views/Body'))
+const Store = lazy(() => import('./Views/DAppStore'))
+const GetHydro = lazy(() => import('./Views/GetHydro'))
+const FinalizeClaim = lazy(() => import('./Views/FinalizeClaim'))
+const Faq = lazy(() => import('./Views/FAQ'))
+const Stats = lazy(() => import('./Views/Stats'))
+const Audits = lazy(() => import('./Views/Audits'))
+const Terms = lazy(() => import('./Views/Terms'))
+const Privacy = lazy(() => import('./Views/Privacy'))
+const About = lazy(() => import('./Views/About'))
+const Contact = lazy(() => import('./Views/Contact'))
+const SubmitDApp = lazy(() => import('./Views/SubmitDApp'))
+const ManageIdentity = lazy(() => import('./Views/ManageYourIdentity'))
+const ManageWallet = lazy(() => import('./Views/DAppStoreWallet'))
+const ManageDApps = lazy(() => import('./Views/DAppsAdded'))
+const DAppCategories = lazy(() => import('./Views/DAppCategories'))
+const DAppDetails = lazy(() => import('./Views/DAppDetails'))
 
 const routeNames = ['/', '/dapp-store', '/get-hydro', '/claim-address', '/faq']
 const MyTabs = withRouter(({ location, children, ...rest }) => {
