+++ src/components/Views/DAppsAdded.jsx
