+++ src/components/Views/Privacy.jsx
