+++ src/components/Views/SnowflakeResolvers.jsx
