+++ src/components/Views/Terms.jsx
